Fix PHP placement, update README, clean up more code.
