+++ .htaccess
@@ -1,0 +1,8 @@
+RewriteEngine On
+RewriteBase /assemblytics/
+
+RewriteCond %{THE_REQUEST} /public/([^\s?]*) [NC]
+RewriteRule ^ %1 [L,NE,R=302]
+
+RewriteRule ^((?!public/).*)$ public/$1 [L,NC]
+
